webhook/requirements.txt
fastapi
uvicorn[standard]
httpx
jinja2

# Email processor dependencies
google-generativeai>=0.8.0
apscheduler>=3.10.0

# Email staging dependencies
psycopg2-binary>=2.9.0
minio>=7.2.0
